# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ledger of Decisions API",
        "description": "決策驅動的記帳 API - 每筆消費都標註決策意圖",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "本地開發環境"
        }
    ],
    "paths": {
        "/expenses/{expense_id}/decision": {
            "get": {
                "tags": [
                    "Decisions"
                ],
                "summary": "查詢決策標註",
                "operationId": "af326e159988e9df7038b86b84e08d08",
                "parameters": [
                    {
                        "name": "expense_id",
                        "in": "path",
                        "description": "消費記錄 ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    },
                    "404": {
                        "description": "消費記錄或決策標註不存在"
                    }
                }
            },
            "put": {
                "tags": [
                    "Decisions"
                ],
                "summary": "更新決策標註",
                "operationId": "b06a4eae5103d383dbb5170caa479a9a",
                "parameters": [
                    {
                        "name": "expense_id",
                        "in": "path",
                        "description": "消費記錄 ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "intent"
                                ],
                                "properties": {
                                    "intent": {
                                        "$ref": "#/components/schemas/Intent"
                                    },
                                    "confidence_level": {
                                        "$ref": "#/components/schemas/ConfidenceLevel"
                                    },
                                    "decision_note": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "更新成功"
                    },
                    "404": {
                        "description": "消費記錄或決策標註不存在"
                    },
                    "422": {
                        "description": "驗證失敗"
                    }
                }
            },
            "post": {
                "tags": [
                    "Decisions"
                ],
                "summary": "新增決策標註",
                "description": "為消費記錄新增決策標註，每筆消費只能有一個決策",
                "operationId": "8c027cd0b15a80d12a5a07bcbda9d57d",
                "parameters": [
                    {
                        "name": "expense_id",
                        "in": "path",
                        "description": "消費記錄 ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "intent"
                                ],
                                "properties": {
                                    "intent": {
                                        "$ref": "#/components/schemas/Intent"
                                    },
                                    "confidence_level": {
                                        "$ref": "#/components/schemas/ConfidenceLevel"
                                    },
                                    "decision_note": {
                                        "type": "string",
                                        "example": "這是必要的開支"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "建立成功"
                    },
                    "404": {
                        "description": "消費記錄不存在"
                    },
                    "422": {
                        "description": "已有決策標註或驗證失敗"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Decisions"
                ],
                "summary": "刪除決策標註",
                "operationId": "96dda031e43e62711f62234fbb9e596e",
                "parameters": [
                    {
                        "name": "expense_id",
                        "in": "path",
                        "description": "消費記錄 ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "刪除成功"
                    },
                    "404": {
                        "description": "消費記錄或決策標註不存在"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Intent": {
                "description": "決策意圖：necessity=必要性, efficiency=效率, enjoyment=享受, recovery=恢復, impulse=衝動",
                "type": "string",
                "enum": [
                    "necessity",
                    "efficiency",
                    "enjoyment",
                    "recovery",
                    "impulse"
                ]
            },
            "ConfidenceLevel": {
                "description": "信心程度：high=高, medium=中, low=低",
                "type": "string",
                "enum": [
                    "high",
                    "medium",
                    "low"
                ]
            },
...
        }
    },
...
}